refactor types and some utils
