+++ src/v3/ontimev3.ts
@@ -6,16 +6,19 @@ import {
 } from '@companion-module/base'
 import { OnTimeInstance, OntimeClient } from '..'
 
+import { connect, disconnectSocket } from './connection'
+import { CustomFields, OntimeEvent } from './ontime-types'
+import { stateobj } from './state'
+
 import { actions } from './actions/index'
 import { feedbacks } from './feedbacks/index'
 import { presets } from './presets'
 import { variables } from './variables'
-import { connect, disconnectSocket } from './connection'
-import { SimpleOntimeEvent, stateobj } from './state'
 
 export class OntimeV3 implements OntimeClient {
 	instance: OnTimeInstance
-	public events: SimpleOntimeEvent[] = []
+	public events: OntimeEvent[] = []
+	public customFields: CustomFields = {}
 	public state = stateobj
 	public log
